refactor(compiler-cli): move typecheck test utils into its own package
This commit moves the test utils used in the typechecking tests into its
own package. This makes them available to be used in the tests of a
different package.

Refs #42966
