+++ packages/compiler-cli/src/ngtsc/typecheck/test/span_comments_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {initMockFileSystem} from '../../file_system/testing';
-import {tcb, TestDeclaration} from './test_utils';
+import {tcb, TestDeclaration} from '../testing';
 
 describe('type check blocks diagnostics', () => {
   beforeEach(() => initMockFileSystem('Native'));
