+++ packages/compiler-cli/src/ngtsc/typecheck/test/type_checker__completion_spec.ts
@@ -14,7 +14,7 @@ import {runInEachFileSystem} from '../../file_system/testing';
 import {getTokenAtPosition} from '../../util/src/typescript';
 import {CompletionKind, GlobalCompletion, TemplateTypeChecker, TypeCheckingConfig} from '../api';
 
-import {getClass, setup, TypeCheckingTarget} from './test_utils';
+import {getClass, setup, TypeCheckingTarget} from '../testing';
 
 runInEachFileSystem(() => {
   describe('TemplateTypeChecker.getGlobalCompletions()', () => {
