+++ packages/compiler-cli/src/ngtsc/typecheck/testing/BUILD.bazel
@@ -1,17 +1,14 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
-
-package(default_visibility = ["//visibility:public"])
+load("//tools:defaults.bzl", "ts_library")
 
 ts_library(
-    name = "test_lib",
+    name = "testing",
     testonly = True,
     srcs = glob([
         "**/*.ts",
     ]),
+    visibility = ["//packages/compiler-cli/src/ngtsc:__subpackages__"],
     deps = [
-        "//packages:types",
         "//packages/compiler",
-        "//packages/compiler-cli/src/ngtsc/diagnostics",
         "//packages/compiler-cli/src/ngtsc/file_system",
         "//packages/compiler-cli/src/ngtsc/file_system/testing",
         "//packages/compiler-cli/src/ngtsc/imports",
@@ -21,21 +18,11 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/program_driver",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/scope",
-        "//packages/compiler-cli/src/ngtsc/shims",
         "//packages/compiler-cli/src/ngtsc/testing",
         "//packages/compiler-cli/src/ngtsc/typecheck",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
         "//packages/compiler-cli/src/ngtsc/typecheck/diagnostics",
-        "//packages/compiler-cli/src/ngtsc/typecheck/testing",
         "//packages/compiler-cli/src/ngtsc/util",
         "@npm//typescript",
-    ],
-)
-
-jasmine_node_test(
-    name = "test",
-    bootstrap = ["//tools/testing:node_no_angular_es5"],
-    deps = [
-        ":test_lib",
     ],
 )
